move dto classes to dto package. restclient classes uses only dto classes.
