+++ src/main/java/br/edu/ifrn/conta/controller/ContaDebitoController.java
@@ -1,5 +1,7 @@
 package br.edu.ifrn.conta.controller;
 
+import br.edu.ifrn.conta.dto.ContaDebitoDTO;
+import br.edu.ifrn.conta.dto.CategoriaDTO;
 import br.edu.ifrn.conta.domain.ContaDebito;
 import br.edu.ifrn.conta.service.CategoriaService;
 import br.edu.ifrn.conta.service.ContaDebitoService;
