+++ src/main/java/br/edu/ifrn/conta/dto/ContaDebitoDTO.java
@@ -1,5 +1,6 @@
-package br.edu.ifrn.conta.controller;
+package br.edu.ifrn.conta.dto;
 
+import br.edu.ifrn.conta.controller.ContaDTO;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
