+++ src/main/java/br/edu/ifrn/conta/dto/LancamentoDTO.java
@@ -1,5 +1,6 @@
-package br.edu.ifrn.conta.controller;
+package br.edu.ifrn.conta.dto;
 
+import br.edu.ifrn.conta.controller.ContaDTO;
 import java.io.Serializable;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
