+++ src/main/java/br/edu/ifrn/conta/restclient/CategoriaRestClient.java
@@ -1,6 +1,6 @@
 package br.edu.ifrn.conta.restclient;
 
-import br.edu.ifrn.conta.controller.CategoriaDTO;
+import br.edu.ifrn.conta.dto.CategoriaDTO;
 import java.util.Optional;
 import lombok.experimental.SuperBuilder;
 import lombok.Data;
