+++ src/main/java/br/edu/ifrn/conta/restclient/TransferenciaRestClient.java
@@ -1,6 +1,6 @@
 package br.edu.ifrn.conta.restclient;
 
-import br.edu.ifrn.conta.controller.TransferenciaDTO;
+import br.edu.ifrn.conta.dto.TransferenciaDTO;
 import java.math.BigDecimal;
 import lombok.experimental.SuperBuilder;
 import lombok.Data;
