+++ src/test/java/br/edu/ifrn/conta/restclient/CategoriaRestClientIT.java
@@ -1,7 +1,7 @@
 package br.edu.ifrn.conta.restclient;
 
 import br.edu.ifrn.conta.ContaApplication;
-import br.edu.ifrn.conta.controller.CategoriaDTO;
+import br.edu.ifrn.conta.dto.CategoriaDTO;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
