+++ src/test/java/br/edu/ifrn/conta/restclient/ContaPatrimonioRestClientFactory.java
@@ -1,7 +1,7 @@
 package br.edu.ifrn.conta.restclient;
 
-import br.edu.ifrn.conta.controller.CategoriaDTO;
-import br.edu.ifrn.conta.controller.ContaPatrimonioDTO;
+import br.edu.ifrn.conta.dto.CategoriaDTO;
+import br.edu.ifrn.conta.dto.ContaPatrimonioDTO;
 import java.util.Optional;
 import lombok.experimental.SuperBuilder;
 import lombok.Data;
