+++ src/test/java/br/edu/ifrn/conta/restclient/LancamentoRestClientFactory.java
@@ -1,6 +1,6 @@
 package br.edu.ifrn.conta.restclient;
 
-import br.edu.ifrn.conta.controller.LancamentoDTO;
+import br.edu.ifrn.conta.dto.LancamentoDTO;
 import lombok.experimental.SuperBuilder;
 import lombok.Data;
