+++ src/test/java/br/edu/ifrn/conta/restclient/SaldoRestClientIT.java
@@ -1,12 +1,12 @@
 package br.edu.ifrn.conta.restclient;
 
 import br.edu.ifrn.conta.ContaApplication;
-import br.edu.ifrn.conta.controller.ContaCreditoDTO;
-import br.edu.ifrn.conta.controller.ContaDebitoDTO;
-import br.edu.ifrn.conta.controller.ContaPatrimonioDTO;
-import br.edu.ifrn.conta.controller.DonoDTO;
-import br.edu.ifrn.conta.controller.LancamentoDTO;
-import br.edu.ifrn.conta.controller.ValorInicialDoDonoNaContaPatrimonioDTO;
+import br.edu.ifrn.conta.dto.ContaCreditoDTO;
+import br.edu.ifrn.conta.dto.ContaDebitoDTO;
+import br.edu.ifrn.conta.dto.ContaPatrimonioDTO;
+import br.edu.ifrn.conta.dto.DonoDTO;
+import br.edu.ifrn.conta.dto.LancamentoDTO;
+import br.edu.ifrn.conta.dto.ValorInicialDoDonoNaContaPatrimonioDTO;
 import static br.edu.ifrn.conta.util.CategoriaConstants.BANCO;
 import static br.edu.ifrn.conta.util.CategoriaConstants.SALARIO;
 import static br.edu.ifrn.conta.util.CategoriaConstants.TRANSPORTE;
