+++ src/test/java/br/edu/ifrn/conta/restclient/TransferenciaRestClientIT.java
@@ -1,9 +1,9 @@
 package br.edu.ifrn.conta.restclient;
 
 import br.edu.ifrn.conta.ContaApplication;
-import br.edu.ifrn.conta.controller.ContaPatrimonioDTO;
-import br.edu.ifrn.conta.controller.DonoDTO;
-import br.edu.ifrn.conta.controller.TransferenciaDTO;
+import br.edu.ifrn.conta.dto.ContaPatrimonioDTO;
+import br.edu.ifrn.conta.dto.DonoDTO;
+import br.edu.ifrn.conta.dto.TransferenciaDTO;
 import static br.edu.ifrn.conta.util.CategoriaConstants.BANCO;
 import static br.edu.ifrn.conta.util.ContaPatrimonioConstants.CONTA_CORRENTE;
 import static br.edu.ifrn.conta.util.ContaPatrimonioConstants.CONTA_INVESTIMENTO;
